Fix some checks
